Update script.js
We refined the script.js file to address an issue where the personality profile charts (main factors and facets) were not rendering on the results page. The primary change involved modifying how color configurations were passed to the Chart.js library within the renderMainFactorBarChart and renderFacetBarChart functions. Specifically, we removed the usage of the potentially problematic Chart.helpers.color utility and switched to directly using the factorColors array and factorConfig.color variables for the chart datasets' borderColor and backgroundColor properties. This was done to resolve a TypeError that was occurring within the Chart.js library, likely due to changes in how color helpers function in the version of Chart.js being loaded from the CDN.